+++ CHANGELOG.md
@@ -11,6 +11,8 @@
   `unsafe` methods for initializing from a raw pointer.
 - Added `CStr16::as_slice_with_nul` to provide immutable access to the
   underlying slice.
+- Added `LoadedImage::load_options_as_bytes` and
+  `LoadedImage::load_options_as_cstr16`.
 
 ### Changed
 
@@ -19,6 +21,8 @@
 - `FileInfo::new`, `FileSystemInfo::new`, and
   `FileSystemVolumeLabel::new` now take their `name` parameter as
   `&CStr16` instead of `&str`, avoiding an implicit string conversion.
+- `LoadImage::set_load_options` now takes a `u8` pointer instead of
+  `Char16`.
 
 ### Removed
 
@@ -28,6 +32,9 @@
 - Removed `FileInfoCreationError::InvalidChar`. This error type is no
   longer needed due to the removal of implicit string conversions in
   file info types.
+- Removed `LoadedImage::load_options`, use
+  `LoadedImage::load_options_as_bytes` or
+  `LoadedImage::load_options_as_cstr16` instead.
 
 ### Fixed